Add memoized change handler for bindState
